+++ common/src/main/kotlin/elements/SerializableWay.kt
@@ -1,11 +1,9 @@
-package api
+package dev.voroscsoki.stratopolis.common.elements
 
 import de.topobyte.osm4j.core.model.iface.OsmNode
 import de.topobyte.osm4j.core.model.iface.OsmWay
-import dev.voroscsoki.stratopolis.common.api.SerializableNode
-import dev.voroscsoki.stratopolis.common.api.SerializableTag
-import dev.voroscsoki.stratopolis.common.api.nodeIds
-import dev.voroscsoki.stratopolis.common.api.tags
+import dev.voroscsoki.stratopolis.common.util.nodeIds
+import dev.voroscsoki.stratopolis.common.util.tags
 import kotlinx.serialization.Serializable
 
 @Serializable
