+++ server/src/main/kotlin/osm/OsmUtils.kt
@@ -3,9 +3,9 @@ package dev.voroscsoki.stratopolis.server.osm
 import de.topobyte.osm4j.core.model.iface.OsmEntity
 import de.topobyte.osm4j.core.model.iface.OsmNode
 import de.topobyte.osm4j.core.model.iface.OsmWay
-import dev.voroscsoki.stratopolis.common.api.Vec3
-import dev.voroscsoki.stratopolis.common.api.nodeIds
-import dev.voroscsoki.stratopolis.common.api.tags
+import dev.voroscsoki.stratopolis.common.util.Vec3
+import dev.voroscsoki.stratopolis.common.util.nodeIds
+import dev.voroscsoki.stratopolis.common.util.tags
 
 fun OsmEntity.isBuilding(): Boolean {
     return this.tags.any { it.key == "building" }
